Migrate input-otp from /docs/components/ to /components/
Port all examples (Basic, Pattern, Form) from the old documentation page
to the new reference page, remove the old page and its route, update
sidebar and home page links, and rewrite E2E tests for the new URL.
