+++ site/ui/e2e/input-otp.spec.ts
@@ -1,107 +1,17 @@
 import { test, expect } from '@playwright/test'
 
-test.describe('Input OTP Documentation Page', () => {
+test.describe('Input OTP Reference Page', () => {
   test.beforeEach(async ({ page }) => {
-    await page.goto('/docs/components/input-otp')
+    await page.goto('/components/input-otp')
   })
 
-  test.describe('Preview', () => {
-    test('renders 6 OTP slots with separator', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const slots = section.locator('[data-slot="input-otp-slot"]')
-      await expect(slots).toHaveCount(6)
-    })
-
-    test('renders separator between groups', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const separator = section.locator('[data-slot="input-otp-separator"]')
-      await expect(separator).toHaveCount(1)
-    })
-
-    test('typing digits fills slots', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const input = section.locator('input[data-otp-input]')
-
-      await input.focus()
-      await input.pressSequentially('123')
-
-      const slots = section.locator('[data-slot="input-otp-slot"]')
-      await expect(slots.nth(0).locator('[data-otp-char]')).toHaveText('1')
-      await expect(slots.nth(1).locator('[data-otp-char]')).toHaveText('2')
-      await expect(slots.nth(2).locator('[data-otp-char]')).toHaveText('3')
-    })
-
-    test('non-digit characters are rejected', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const input = section.locator('input[data-otp-input]')
-
-      await input.focus()
-      await input.pressSequentially('1a2b3')
-
-      const slots = section.locator('[data-slot="input-otp-slot"]')
-      await expect(slots.nth(0).locator('[data-otp-char]')).toHaveText('1')
-      await expect(slots.nth(1).locator('[data-otp-char]')).toHaveText('2')
-      await expect(slots.nth(2).locator('[data-otp-char]')).toHaveText('3')
-    })
-
-    test('backspace removes last character', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const input = section.locator('input[data-otp-input]')
-
-      await input.focus()
-      await input.pressSequentially('12')
-      await input.press('Backspace')
-
-      const slots = section.locator('[data-slot="input-otp-slot"]')
-      await expect(slots.nth(0).locator('[data-otp-char]')).toHaveText('1')
-      await expect(slots.nth(1).locator('[data-otp-char]')).toHaveText('')
-    })
-
-    test('paste fills multiple slots', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const input = section.locator('input[data-otp-input]')
-
-      await input.focus()
-      // Use clipboard API to paste
-      await page.evaluate(() => {
-        const input = document.querySelector('input[data-otp-input]') as HTMLInputElement
-        input.focus()
-        const pasteEvent = new ClipboardEvent('paste', {
-          clipboardData: new DataTransfer(),
-        })
-        pasteEvent.clipboardData!.setData('text', '456789')
-        input.dispatchEvent(pasteEvent)
-      })
-
-      const slots = section.locator('[data-slot="input-otp-slot"]')
-      await expect(slots.nth(0).locator('[data-otp-char]')).toHaveText('4')
-      await expect(slots.nth(1).locator('[data-otp-char]')).toHaveText('5')
-      await expect(slots.nth(2).locator('[data-otp-char]')).toHaveText('6')
-      await expect(slots.nth(3).locator('[data-otp-char]')).toHaveText('7')
-      await expect(slots.nth(4).locator('[data-otp-char]')).toHaveText('8')
-      await expect(slots.nth(5).locator('[data-otp-char]')).toHaveText('9')
-    })
-
-    test('input truncates at maxLength', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const input = section.locator('input[data-otp-input]')
-
-      await input.focus()
-      await input.pressSequentially('12345678')
-
-      // Only first 6 chars should be displayed
-      const slots = section.locator('[data-slot="input-otp-slot"]')
-      await expect(slots.nth(5).locator('[data-otp-char]')).toHaveText('6')
-    })
-
-    test('active slot has data-active=true', async ({ page }) => {
-      const section = page.locator('[bf-s^="InputOTPPreviewDemo_"]:not([data-slot])').first()
-      const input = section.locator('input[data-otp-input]')
-
-      await input.focus()
-
-      const slots = section.locator('[data-slot="input-otp-slot"]')
-      await expect(slots.nth(0)).toHaveAttribute('data-active', 'true')
+  test.describe('Preview (Playground)', () => {
+    test('renders OTP input in playground', async ({ page }) => {
+      const preview = page.locator('[data-input-otp-preview]').first()
+      await expect(preview).toBeVisible()
+      // Playground renders InputOTP with hidden native input
+      const input = preview.locator('input')
+      await expect(input.first()).toBeAttached()
     })
   })
 
@@ -158,9 +68,10 @@ test.describe('Input OTP Documentation Page', () => {
       // does not trigger onclick property handlers set by hydration
       const result = await page.evaluate(() => {
         const section = document.querySelector('[bf-s^="InputOTPFormDemo_"]:not([data-slot])') as HTMLElement
-        const btn = section?.querySelector('button[bf="s11"]') as HTMLButtonElement
-        btn?.click()
-        return btn?.textContent
+        const buttons = section?.querySelectorAll('button') as NodeListOf<HTMLButtonElement>
+        const verifyBtn = Array.from(buttons).find(b => b.textContent?.includes('Verify'))
+        verifyBtn?.click()
+        return verifyBtn?.textContent
       })
       expect(result).toBe('Verifying...')
 
@@ -176,8 +87,9 @@ test.describe('Input OTP Documentation Page', () => {
 
       await page.evaluate(() => {
         const section = document.querySelector('[bf-s^="InputOTPFormDemo_"]:not([data-slot])') as HTMLElement
-        const btn = section?.querySelector('button[bf="s11"]') as HTMLButtonElement
-        btn?.click()
+        const buttons = section?.querySelectorAll('button') as NodeListOf<HTMLButtonElement>
+        const verifyBtn = Array.from(buttons).find(b => b.textContent?.includes('Verify'))
+        verifyBtn?.click()
       })
 
       await expect(section.locator('text=Invalid code. Please try again.')).toBeVisible({ timeout: 5000 })
