+++ site/ui/pages/components/input-otp.tsx
@@ -8,6 +8,11 @@
 import { InputOTPPlayground } from '@/components/input-otp-playground'
 import { InputOTPUsageDemo } from '@/components/input-otp-usage-demo'
 import {
+  InputOTPBasicDemo,
+  InputOTPPatternDemo,
+  InputOTPFormDemo,
+} from '@/components/input-otp-demo'
+import {
   DocPage,
   PageHeader,
   Section,
@@ -23,6 +28,10 @@ const tocItems: TocItem[] = [
   { id: 'preview', title: 'Preview' },
   { id: 'installation', title: 'Installation' },
   { id: 'usage', title: 'Usage' },
+  { id: 'examples', title: 'Examples' },
+  { id: 'basic', title: 'Basic', branch: 'start' },
+  { id: 'pattern', title: 'Pattern', branch: 'child' },
+  { id: 'form', title: 'Form', branch: 'end' },
   { id: 'api-reference', title: 'API Reference' },
 ]
 
@@ -58,6 +67,133 @@ function InputOTPDemo() {
         <InputOTPSlot index={5} />
       </InputOTPGroup>
     </InputOTP>
+  )
+}`
+
+const basicCode = `import {
+  InputOTP,
+  InputOTPGroup,
+  InputOTPSlot,
+} from "@/components/ui/input-otp"
+
+export function InputOTPBasicDemo() {
+  return (
+    <InputOTP maxLength={4}>
+      <InputOTPGroup>
+        <InputOTPSlot index={0} />
+        <InputOTPSlot index={1} />
+        <InputOTPSlot index={2} />
+        <InputOTPSlot index={3} />
+      </InputOTPGroup>
+    </InputOTP>
+  )
+}`
+
+const patternCode = `import {
+  InputOTP,
+  InputOTPGroup,
+  InputOTPSlot,
+  REGEXP_ONLY_DIGITS_AND_CHARS,
+} from "@/components/ui/input-otp"
+
+export function InputOTPPatternDemo() {
+  return (
+    <div className="space-y-2">
+      <InputOTP maxLength={6} pattern={REGEXP_ONLY_DIGITS_AND_CHARS}>
+        <InputOTPGroup>
+          <InputOTPSlot index={0} />
+          <InputOTPSlot index={1} />
+          <InputOTPSlot index={2} />
+          <InputOTPSlot index={3} />
+          <InputOTPSlot index={4} />
+          <InputOTPSlot index={5} />
+        </InputOTPGroup>
+      </InputOTP>
+      <p className="text-sm text-muted-foreground">
+        Accepts letters and numbers.
+      </p>
+    </div>
+  )
+}`
+
+const formCode = `"use client"
+
+import { createSignal, createMemo, onCleanup } from "@barefootjs/dom"
+import {
+  InputOTP,
+  InputOTPGroup,
+  InputOTPSlot,
+  InputOTPSeparator,
+} from "@/components/ui/input-otp"
+
+export function InputOTPFormDemo() {
+  const [value, setValue] = createSignal('')
+  const [status, setStatus] = createSignal('idle')
+  const [canResend, setCanResend] = createSignal(true)
+  const [countdown, setCountdown] = createSignal(0)
+
+  const isComplete = createMemo(() => value().length === 6)
+
+  const handleSubmit = () => {
+    if (!isComplete()) return
+    setStatus('loading')
+    setTimeout(() => {
+      if (value() === '123456') {
+        setStatus('success')
+      } else {
+        setStatus('error')
+      }
+    }, 1500)
+  }
+
+  const handleResend = () => {
+    if (!canResend()) return
+    setCanResend(false)
+    setCountdown(30)
+    setValue('')
+    setStatus('idle')
+
+    const timer = setInterval(() => {
+      setCountdown(prev => {
+        if (prev <= 1) {
+          clearInterval(timer)
+          setCanResend(true)
+          return 0
+        }
+        return prev - 1
+      })
+    }, 1000)
+    onCleanup(() => clearInterval(timer))
+  }
+
+  return (
+    <div className="space-y-4">
+      <InputOTP maxLength={6} value={value()} onValueChange={setValue}
+        disabled={status() === 'loading' || status() === 'success'}>
+        <InputOTPGroup>
+          <InputOTPSlot index={0} />
+          <InputOTPSlot index={1} />
+          <InputOTPSlot index={2} />
+        </InputOTPGroup>
+        <InputOTPSeparator />
+        <InputOTPGroup>
+          <InputOTPSlot index={3} />
+          <InputOTPSlot index={4} />
+          <InputOTPSlot index={5} />
+        </InputOTPGroup>
+      </InputOTP>
+      <div className="flex items-center gap-3">
+        <button disabled={!isComplete() || status() === 'loading'}
+          onClick={handleSubmit}>
+          {status() === 'loading' ? 'Verifying...' : 'Verify'}
+        </button>
+        <button disabled={!canResend()} onClick={handleResend}>
+          {canResend() ? 'Resend code' : \`Resend in \${countdown()}s\`}
+        </button>
+      </div>
+      {status() === 'success' && <p>Code verified successfully!</p>}
+      {status() === 'error' && <p>Invalid code. Please try again.</p>}
+    </div>
   )
 }`
 
@@ -125,6 +261,23 @@ export function InputOTPRefPage() {
           <Example title="" code={usageCode}>
             <InputOTPUsageDemo />
           </Example>
+        </Section>
+
+        {/* Examples */}
+        <Section id="examples" title="Examples">
+          <div className="space-y-8">
+            <Example title="Basic" code={basicCode}>
+              <InputOTPBasicDemo />
+            </Example>
+
+            <Example title="Pattern" code={patternCode}>
+              <InputOTPPatternDemo />
+            </Example>
+
+            <Example title="Form" code={formCode}>
+              <InputOTPFormDemo />
+            </Example>
+          </div>
         </Section>
 
         {/* API Reference */}
